+++ lingui.config.ts
@@ -16,6 +16,14 @@ const config: LinguiConfig = {
       include: ['packages/select/index.js'],
       path: 'packages/select/locales/{locale}/messages',
     },
+    {
+      include: ['packages/attention/index.js'],
+      path: 'packages/attention/locales/{locale}/messages',
+    },
+    {
+      include: ['packages/expandable/index.js'],
+      path: 'packages/expandable/locales/{locale}/messages',
+    },
   ],
   compileNamespace: 'es',
   sourceLocale: 'en',