+++ packages/affix/index.js
@@ -1,13 +1,14 @@
-import { html, LitElement, css } from 'lit';
+import { html } from 'lit';
+import WarpElement from '@warp-ds/elements-core';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { suffix, prefix } from '@warp-ds/css/component-classes';
+import '@warp-ds/icons/elements/search-16'
+import '@warp-ds/icons/elements/close-16'
 
 import { fclasses } from '../utils';
 
-class WarpAffix extends LitElement {
-  static styles = css`
-    @unocss-placeholder
-  `;
+class WarpAffix extends WarpElement {
+  static styles = [WarpElement.styles];
 
   static properties = {
     ariaLabel: { type: String, attribute: 'aria-label' },