+++ packages/box/index.js
@@ -1,29 +1,32 @@
-import { css, html, LitElement } from 'lit';
+import { css, html, nothing } from 'lit';
+import WarpElement from "@warp-ds/elements-core";
 import { fclasses } from '../utils';
 import { box as ccBox } from '@warp-ds/css/component-classes';
 
-class WarpBox extends LitElement {
+class WarpBox extends WarpElement {
   static properties = {
     bleed: { type: Boolean },
     bordered: { type: Boolean },
     info: { type: Boolean },
     neutral: { type: Boolean },
+    role: { type: String },
   };
 
   // Slotted elements remain in lightDOM which allows for control of their style outside of shadowDOM.
   // ::slotted([Simple Selector]) confirms to Specificity rules, but (being simple) does not add weight to lightDOM skin selectors,
   // so never gets higher Specificity. Thus in order to overwrite style linked within shadowDOM, we need to use !important.
   // https://stackoverflow.com/a/61631668
-  static styles = 
+  static styles = [
+    WarpElement.styles,
     css`
-      @unocss-placeholder
       :host {
-          display: block;
-        }
-        ::slotted(:last-child) {
-          margin-bottom: 0 !important;
-        }
-      `;
+        display: block;
+      }
+      ::slotted(:last-child) {
+        margin-bottom: 0 !important;
+      }
+    `,
+  ];
 
   get _class() {
     return fclasses({
@@ -35,9 +38,13 @@ class WarpBox extends LitElement {
     });
   }
 
+  get _optOutRoleWithDefault() {
+    return this.role === '' ? nothing : this.role ?? 'region'
+  }
+
   render() {
     return html`
-      <div class="${this._class}">
+      <div role="${this._optOutRoleWithDefault}" class="${this._class}">
         <slot></slot>
       </div>
     `;