+++ packages/expandable/index.js
@@ -1,12 +1,19 @@
-import { css, html, LitElement } from 'lit';
-import { fclasses, kebabCaseAttributes } from '../utils';
+import { css, html } from 'lit'
+import WarpElement from '@warp-ds/elements-core'
+import { fclasses, kebabCaseAttributes } from '../utils'
 import {
   box as ccBox,
   expandable as ccExpandable,
-} from '@warp-ds/css/component-classes';
-import { ifDefined } from 'lit/directives/if-defined.js';
+} from '@warp-ds/css/component-classes'
 
-class WarpExpandable extends kebabCaseAttributes(LitElement) {
+import { ifDefined } from 'lit/directives/if-defined.js'
+import { i18n } from '@lingui/core'
+import { messages as enMessages } from './locales/en/messages.mjs'
+import { messages as nbMessages } from './locales/nb/messages.mjs'
+import { messages as fiMessages } from './locales/fi/messages.mjs'
+import { activateI18n } from '../i18n'
+
+class WarpExpandable extends kebabCaseAttributes(WarpElement) {
   static properties = {
     expanded: { type: Boolean, reflect: true },
     title: { type: String },
@@ -19,18 +26,30 @@ class WarpExpandable extends kebabCaseAttributes(LitElement) {
     animated: { type: Boolean },
     headingLevel: { type: Number },
     _hasTitle: { type: Boolean, state: true },
-  };
+    _showChevronUp: { type: Boolean, state: true },
+  }
 
   constructor() {
-    super();
+    super()
+    activateI18n(enMessages, nbMessages, fiMessages)
 
-    this.expanded = false;
-    this.animated = false;
-    this.info = false;
-    this.box = false;
-    this.bleed = false;
-    this.noChevron = false;
-    this._hasTitle = true;
+    this.expanded = false
+    this.animated = false
+    this.info = false
+    this.box = false
+    this.bleed = false
+    this.noChevron = false
+    this._hasTitle = true
+    this._showChevronUp = this.expanded
+  }
+
+  updated(changedProperties) {
+    // We need a slight delay for the animation since it has a transition-duration of 150ms:
+    if (changedProperties.has('expanded')) {
+      setTimeout(() => {
+        this._showChevronUp = this.expanded
+      }, 200)
+    }
   }
 
   // Slotted elements remain in lightDOM which allows for control of their style outside of shadowDOM.
@@ -38,8 +57,8 @@ class WarpExpandable extends kebabCaseAttributes(LitElement) {
   // so never gets higher Specificity. Thus in order to overwrite style linked within shadowDOM, we need to use !important.
   // https://stackoverflow.com/a/61631668
   static styles = [
+    WarpElement.styles,
     css`
-      @unocss-placeholder
       :host {
         display: block;
       }
@@ -47,13 +66,13 @@ class WarpExpandable extends kebabCaseAttributes(LitElement) {
         margin-bottom: 0px !important;
       }
     `,
-  ];
+  ]
 
   firstUpdated() {
     this._hasTitle =
       !!this.title ||
       this.renderRoot.querySelector("slot[name='title']").assignedNodes()
-        .length > 0;
+        .length > 0
   }
 
   get _expandableSlot() {
@@ -65,10 +84,43 @@ class WarpExpandable extends kebabCaseAttributes(LitElement) {
       })}
     >
       <slot></slot>
-    </div>`;
+    </div>`
+  }
+
+  get _chevronUpClasses() {
+    return fclasses({
+      [ccExpandable.chevronTransform]: true,
+      [ccExpandable.chevronCollapse]: !this.expanded && this._showChevronUp,
+    })
+  }
+
+  get _chevronDownClasses() {
+    return fclasses({
+      [ccExpandable.chevronTransform]: true,
+      [ccExpandable.chevronExpand]: this.expanded && !this._showChevronUp,
+    })
+  }
+
+  chevronDownTitle() {
+    return i18n._({
+      id: 'icon.title.chevron-down',
+      message: 'Downward arrow',
+      comment:
+        'Default screenreader message for chevron down icon in the expandable component',
+    })
+  }
+
+  chevronUpTitle() {
+    return i18n._({
+      id: 'icon.title.chevron-up',
+      message: 'Upward arrow',
+      comment:
+        'Default screenreader message for chevron up icon in the expandable component',
+    })
   }
 
   render() {
+    // Will keep the inline-svg:s for expandable due to issues with setting css-classes on an element inside the shadow DOM:
     return html` <div
       class=${fclasses({
         [ccExpandable.expandable]: true,
@@ -90,19 +142,56 @@ class WarpExpandable extends kebabCaseAttributes(LitElement) {
             >
               <div class="${ccExpandable.title}">
                 ${this.title
-                  ? html`<span class="${ccExpandable.titleType}">${this.title}</span>`
+                  ? html`<span class="${ccExpandable.titleType}"
+                      >${this.title}</span
+                    >`
                   : html`<slot name="title"></slot>`}
                 ${this.noChevron
                   ? ''
                   : html`<div
                       class=${fclasses({
                         [ccExpandable.chevron]: true,
-                        [ccExpandable.chevronExpanded]: this.expanded,
                         [ccExpandable.chevronBox]: this.box,
                         [ccExpandable.chevronNonBox]: !this.box,
                       })}
                     >
-                      <w-icon-chevron-down-16></w-icon-chevron-down-16>
+                      ${this._showChevronUp
+                        ? html`<svg
+                            id="chevron-up"
+                            class=${this._chevronUpClasses}
+                            xmlns="http://www.w3.org/2000/svg"
+                            width="16"
+                            height="16"
+                            fill="none"
+                            viewBox="0 0 16 16"
+                          >
+                            <title>${this.chevronUpTitle()}</title>
+                            <path
+                              stroke="currentColor"
+                              stroke-linecap="round"
+                              stroke-linejoin="round"
+                              stroke-width="1.5"
+                              d="M2.5 11 8 5.5l5.5 5.5"
+                            ></path>
+                          </svg>`
+                        : html`<svg
+                            id="chevron-down"
+                            class=${this._chevronDownClasses}
+                            xmlns="http://www.w3.org/2000/svg"
+                            width="16"
+                            height="16"
+                            fill="none"
+                            viewBox="0 0 16 16"
+                          >
+                            <title>${this.chevronDownTitle()}</title>
+                            <path
+                              stroke="currentColor"
+                              stroke-linecap="round"
+                              stroke-linejoin="round"
+                              stroke-width="1.5"
+                              d="M2.5 5.5 8 11l5.5-5.5"
+                            ></path>
+                          </svg>`}
                     </div>`}
               </div>
             </button>
@@ -121,12 +210,12 @@ class WarpExpandable extends kebabCaseAttributes(LitElement) {
           >
             ${this._expandableSlot}
           </div>`}
-    </div>`;
+    </div>`
   }
 }
 
 if (!customElements.get('w-expandable')) {
-  customElements.define('w-expandable', WarpExpandable);
+  customElements.define('w-expandable', WarpExpandable)
 }
 
-export { WarpExpandable };
+export { WarpExpandable }